feat: final backend.

+++ src/main/java/com/prodev/ecomarket/purchases/domain/model/commands/CreatePurchaseCommand.java
@@ -1,4 +1,5 @@
 package com.prodev.ecomarket.purchases.domain.model.commands;
 
-public record CreatePurchaseCommand (Double totalAmount, String status, String paymentMethod, int customerId, int productId) {
+public record CreatePurchaseCommand (Double totalAmount, String status, String paymentMethod,
+                                        Long customerId, int productId) {
 }
