+++ src/main/java/com/prodev/ecomarket/purchases/interfaces/rest/resources/CreatePurchaseResource.java
@@ -6,6 +6,6 @@ public record CreatePurchaseResource(Date date,
                                      double totalAmount,
                                      String status,
                                      String paymentMethod,
-                                     int customerId,
+                                     Long customerId,
                                      int paymentId) {
 }
